+++ mod-record-specifications-validator/src/main/java/org/folio/rspec/i18n/TranslationProvider.java
@@ -1,0 +1,7 @@
+package org.folio.rspec.i18n;
+
+@FunctionalInterface
+public interface TranslationProvider {
+
+  String format(String key, Object... args);
+}