+++ mod-record-specifications-validator/src/main/java/org/folio/rspec/validation/validator/marc/model/MarcIndicator.java
@@ -1,0 +1,8 @@
+package org.folio.rspec.validation.validator.marc.model;
+
+public record MarcIndicator(Reference reference, Character value) {
+
+  public Integer order() {
+    return reference().getIndicatorIndex();
+  }
+}