+++ mod-record-specifications-validator/src/main/java/org/folio/rspec/validation/validator/marc/model/MarcSubfield.java
@@ -1,0 +1,8 @@
+package org.folio.rspec.validation.validator.marc.model;
+
+public record MarcSubfield(Reference reference, String value) {
+
+  public Character code() {
+    return reference.getSubfield();
+  }
+}